% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add.ribbon}
\alias{add.ribbon}
\title{Plot time series ribbon for posterior draws}
\usage{
add.ribbon(x, z, col, alpha.level, alpha.min = 0, alpha.max = 1)
}
\arguments{
\item{x}{Vector of years}

\item{z}{Matrix of values (years by draws)}

\item{col}{The base color without transparency.}

\item{alpha.level}{This is the credible region, e.g. 0.5 will contain
50% of the posterior draws. You can pass multiple levels to get a
shading effect. The transparaceny is proportional to the credible
region (wider regions are lighter).}

\item{alpha.min}{Minimum level of shading, for fine control of
coloring. Defaults to 0.}

\item{alpha.max}{See alpha.min. Defaults to 1.}
}
\value{
Nothing, addings filled polygon to current plot.
}
\description{
Plot time series ribbon for posterior draws
}
