% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{AgeModel}
\alias{AgeModel}
\title{Stochastic SRA to calculate the trajectory given parameters}
\usage{
AgeModel(Catch, AgeMat, Steep, NatMort, AgeMax, Carry, Weight,
  InitialDeplete, Sigma, AgeVulnOffset = -1, ProcessError = TRUE,
  simulation = NULL, carry_form = "ssb")
}
\arguments{
\item{Catch}{To do}

\item{AgeMat}{Age at maturity}

\item{Steep}{Steepness}

\item{NatMort}{Natural mortality}

\item{AgeMax}{Maximum age, above which there is a plus group.}

\item{Carry}{Carrying capacity.}

\item{Weight}{Vector of weights at age, calculated from growth
parameters a and b.}

\item{InitialDeplete}{Initial level of depletion.}

\item{Sigma}{Process error variance, i.e. N(0,Sigma).}

\item{AgeVulnOffset}{Optional parameter for specifying age at
vulnerability, defaulting to AgeMat-1. See ?run.SIR for more info.}

\item{ProcessError}{A flag for whether to include process error,
defaults to TRUE.}

\item{simulation}{A list with elements 'NYears' and 'FishMort'. If
supplied the model will run a simulation under that level of U. It it
designed to be used to test equilibrium conditions but can be expanded
later for more expansive simulation testing.}
}
\description{
Stochastic SRA to calculate the trajectory given parameters
}
