% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baranov_catches.R
\name{baranov_catches}
\alias{baranov_catches}
\title{Calculate Catches based on Baranov Equation}
\usage{
baranov_catches(log_f, m, sel, b_a, catch, use = 1)
}
\arguments{
\item{log_f}{log fishing mortality}

\item{m}{natural mortality}

\item{sel}{selectivity at age}

\item{b_a}{biomass at age}

\item{catch}{observed catches}

\item{use}{1 = optimzation, anything else equals catches}
}
\value{
catch or ss
}
\description{
Calculate Catches based on Baranov Equation
}
