% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_penalties}
\alias{check_penalties}
\title{Check whether the penalties list is valid.}
\usage{
check_penalties(penalties)
}
\arguments{
\item{penalties}{A list passed by the user}
}
\value{
Nothing, an error occurs if a problem is found.
}
\description{
Check whether the penalties list is valid.
}
