% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cod.R
\docType{data}
\name{cod}
\alias{cod}
\title{Sample data for atlantic cod from RAM}
\format{
A data frame with 53 rows and 50 variables:
\describe{
\item{catch}{catch, metric tons}
\item{b_v_bmsy}{B/Bmsy}
\item{u_v_umsy}{U/Umsy}
\item{b_rel}{depletion}
...
}
}
\source{
\url{http://ramlegacy.org/}
}
\usage{
cod
}
\description{
A dataset containing RAM data and auxillary information for
atlantic cod (stocklong "Atlantic cod IIIa (west) and IV-VIId")
}
\keyword{datasets}
