% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srafit.R
\name{construct_srafit}
\alias{construct_srafit}
\title{Construct an object of class 'srafit'}
\usage{
construct_srafit(fit)
}
\arguments{
\item{fit}{A list}
}
\value{
A list of class 'srafit'
}
\description{
Construct an object of class 'srafit'
}
