% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose-sraplus.R
\name{diagnose_sraplus}
\alias{diagnose_sraplus}
\title{Run minimal diagnostics on sraplus fits}
\usage{
diagnose_sraplus(fit, driors)
}
\arguments{
\item{fit}{a fitted sraplus model}

\item{driors}{the driors used to fit the model}
}
\value{
a list of diagnostics
}
\description{
Run minimal diagnostics on sraplus fits
}
\examples{

diagnose_sraplus(sraplus_fit, sraplus_driors)
}
