% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{draw.priors}
\alias{draw.priors}
\title{Quick function to return priors for the Mexican examples.}
\usage{
draw.priors(N, iStock)
}
\arguments{
\item{N}{Number of draws}

\item{iStock}{Stock number (temporary)}
}
\value{
A data.frame that contains N random draws from all priors
}
\description{
Quick function to return priors for the Mexican examples.
}
