% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ancestors.R
\name{find_ancestors}
\alias{find_ancestors}
\title{Find ancestors Ported fromhttps://github.com/James-Thorson-NOAA/FishLife/blob/main/R/Find_ancestors.R}
\usage{
find_ancestors(
  child_num,
  Database = marlin::FishBase_and_RAM,
  ParentChild_gz = Database$ParentChild_gz
)
}
\arguments{
\item{child_num}{row number of \code{ParentChild_gz} for which to find ancestors}
}
\value{
vector of row numbers of \code{ParentChild_gz} for ancestors (including \code{child_num})
}
\description{
Find higher taxonomic levels for a given taxon (e.g., Class and Order for a given Family)
}
