% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-prior-regressions.R
\name{fit_prior_regressions}
\alias{fit_prior_regressions}
\title{fir-sraplus-regressions}
\usage{
fit_prior_regressions(
  split,
  model_structure,
  use_splits = TRUE,
  chains = 4,
  cores = 4,
  iter = 2000,
  adapt_delta = adapt_delta,
  max_treedepth = max_treedepth,
  produce = "summary",
  refresh = 0,
  post_draws = 10,
  ...
)
}
\arguments{
\item{split}{}

\item{model_structure}{}

\item{use_splits}{}

\item{chains}{}

\item{cores}{}

\item{iter}{}

\item{adapt_delta}{}

\item{max_treedepth}{}

\item{produce}{}

\item{refresh}{}

\item{post_draws}{}

\item{...}{}
}
\value{
list with fitted object
}
\description{
fir-sraplus-regressions
}
