% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-sraplus.R
\name{fit_sraplus}
\alias{fit_sraplus}
\title{Run sraplus}
\usage{
fit_sraplus(
  driors,
  include_fit = TRUE,
  seed = 42,
  plim = 0.05,
  model_name = "sraplus_tmb",
  randos = "log_proc_errors",
  draws = 1e+05,
  n_keep = 2000,
  engine = "sir",
  cores = 4,
  chains = 1,
  cleanup = FALSE,
  max_treedepth = 10,
  adapt_delta = 0.8,
  estimate_shape = FALSE,
  estimate_qslope = FALSE,
  estimate_q = TRUE,
  estimate_proc_error = TRUE,
  estimate_initial_state = TRUE,
  estimate_k = TRUE,
  estimate_f = FALSE,
  learn_rate = 0.001,
  analytical_q = FALSE,
  use_baranov = TRUE,
  include_m = FALSE,
  ci = 0.89,
  try_again = FALSE,
  eps = 1e-06,
  max_time = Inf,
  eval.max = 200,
  iter.max = 150,
  rel.tol = 1e-10,
  loopnum = 2,
  newtonsteps = 0,
  tune_prior_predictive = TRUE,
  index_fit_tuner = "sir",
  refresh = 250,
  log_bias_correct = TRUE,
  workers = 1,
  thin_draws = FALSE,
  thin_rate = 0.5,
  catch_cv = 0.01,
  non_centered = FALSE,
  ...
)
}
\arguments{
\item{driors}{a list of driors passed from sraplus::format_driors}

\item{include_fit}{logical indicating whether to return the fitted object}

\item{seed}{seed for model runs}

\item{plim}{cutoff (in units of B/K) for hockey stick PT function}

\item{model_name}{the name of the sraplus TMB version to be run, defaults to "sraplus_tmb"}

\item{randos}{random effects when passing to TMB}

\item{draws}{the number of SIR samples to run}

\item{n_keep}{the number of SIR samples to keep}

\item{engine}{one of 'sir','stan', or 'tmb'}

\item{cores}{number of cores for stan fits}

\item{chains}{number of chains for stan fits}

\item{cleanup}{logical indicating whether to remove the compiled TMB model after running}

\item{max_treedepth}{max_treedepth for models fit using stan}

\item{adapt_delta}{adap_delta for models fit using stan}

\item{estimate_shape}{logical indicating whether to estimate the shape parameter of the Pella-Tomlinson model.
If FALSE shape parameter is held at the initial value, either default or supplied}

\item{estimate_qslope}{logical indicating whether to estimate a slope parameter for q in CPUE fitting.
If FALSE q_slope is held at the initial value, either default or supplied. if TRUE, \code{estimate_proc_error} should be set to FALSE}

\item{estimate_q}{estimate catchability coefficient or leave fixed}

\item{estimate_proc_error}{logical indicating whether to estimate process errors.
If FALSE process errors are not included in the model}

\item{estimate_initial_state}{}

\item{estimate_k}{estimate carrying capacity}

\item{estimate_f}{estimate fishing mortality}

\item{learn_rate}{learn_rate for HMC}

\item{analytical_q}{use analytical q instead of empirical}

\item{use_baranov}{use baranov catch equation}

\item{include_m}{include natural mortality (deprecated)}

\item{ci}{confidence/credible interval range for summaries}

\item{try_again}{try fit again with adjusted starting guess}

\item{eps}{eps for nlminb}

\item{max_time}{max_time for nlminb}

\item{eval.max}{eval.max for nlminb}

\item{iter.max}{iter.max for nlminb}

\item{rel.tol}{for nlminb}

\item{loopnum}{for nlminb}

\item{newtonsteps}{for TMBhelper}

\item{tune_prior_predictive}{tune prior predictive to resolve borel's paradox}

\item{index_fit_tuner}{"sir" tunes priors for index fit}

\item{refresh}{refresh rate for Stan}

\item{log_bias_correct}{correct for log-transformation bias}

\item{workers}{number of workers for parallel processing of stan outputs}

\item{thin_draws}{true or false thing draws}

\item{thin_rate}{thinning rate}

\item{catch_cv}{assumed CV for catch data when estimating catches rather than treating as data}

\item{non_centered}{use centered or non-centered form for process errors}

\item{...}{additional parameters}
}
\value{
a fitted sraplus object
}
\description{
The main function for taking an object created by \code{format_driors} and fitting a surplus production
model using \code{sraplus}
}
