% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-driors.R
\name{format_driors}
\alias{format_driors}
\title{Format data and priors for sraplus}
\usage{
format_driors(taxa = "gadus morhua", initial_b = 1,
  initial_b_sd = 0.1, terminal_b = NA, terminal_b_sd = 0.1,
  carry = NA, carry_sd = 0.1, u_v_umsy = NA, u_years = NA,
  u_sd = 0.1, final_u = NA, final_u_sd = NA, catch = NA,
  years = NA, index = NA, effort = NA, ref_type = "k",
  index_years = 1, effort_years = 1, use_heuristics = FALSE,
  fmi = c(research = NA, management = NA, enforcement = NA,
  socioeconomics = NA), fmi_sd = NA, sar = NA, sar_sd = NA,
  f_sd = 0.1)
}
\arguments{
\item{taxa}{the genus and species of the species (case insensitive)}

\item{initial_b}{b reference point in the initial year, units of depletion of B/Bmsy (set ref_type accordingly)}

\item{initial_b_sd}{sigma associated with initial b reference point}

\item{terminal_b}{b reference point in the terminal year, units of depletion of B/Bmsy (set ref_type accordingly)}

\item{terminal_b_sd}{sigma associated with terminal b reference point}

\item{carry}{prior on carrying capacity, deprecated}

\item{carry_sd}{cv associated with prior on carrying capacity, deprecated}

\item{u_v_umsy}{u/umsy data over time}

\item{u_years}{years in which u/umsy data are available}

\item{u_sd}{cv associated with u/umsy data}

\item{final_u}{vector of priors on u/umsy in the terminal years}

\item{final_u_sd}{vector of cvs on u/umsy in the terminal years}

\item{catch}{vector of catches over lifetime of fishery}

\item{years}{vector of years that the catch data correspond to}

\item{index}{vector of an abundance index}

\item{effort}{vector of an effort series}

\item{ref_type}{k if initial and final depletions are in depletion units, b if in b/bmsy units}

\item{index_years}{the years in which abundance index data are available}

\item{effort_years}{years in which effort data are available}

\item{use_heuristics}{logical,TRUE uses catch-msy hueristics for priors, FALSE requires user to pass them}

\item{fmi}{named vector of fisheries management index scrores}

\item{fmi_sd}{overwrite fmi prediction sd}

\item{sar}{swept area ratio}

\item{sar_sd}{overwrite sar prediction sd}

\item{f_sd}{deprecated}
}
\value{
a list of data and priors
}
\description{
Format data and priors for sraplus
}
