% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eq_catch.R
\name{get_eq_catch}
\alias{get_eq_catch}
\title{get equilibrium catch}
\usage{
get_eq_catch(log_f, rzeroC, NatMort, maxage, vuln, Weight, alpha, betav,
  biomass = FALSE)
}
\arguments{
\item{log_f}{log fishing mortality}

\item{rzeroC}{corrected recruitment deviate}

\item{NatMort}{natural mortality}

\item{maxage}{max age}

\item{vuln}{fishing vulnerability at age}

\item{Weight}{weight at age}

\item{alpha}{recruitment alpha}

\item{betav}{recruitment beta}

\item{biomass}{logical TRUE = Bmsy, false = catch}
}
\value{
either biomass or catch
}
\description{
get equilibrium catch
}
