% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_prior}
\alias{get_prior}
\title{Extract prior information for initial biomass, terminal U/UMSY (ustatus)
and terminal B/BMSY (bstatus) from a fit, which is useful for plotting
and other tasks.}
\usage{
get_prior(fit, metric, interval = TRUE, percentile = 0.95,
  n.points = 1000)
}
\arguments{
\item{fit}{An object of class srafit}

\item{metric}{Which metric to use from 'carry', 'ustatus', 'bstatus'.}

\item{interval}{Whether to return the prior credible interval or a
sequence of points evaluated at the prior density.}

\item{percentile}{The interval converage (e.g., .95 is the 95\% credible
interval)}

\item{n.points}{The number of points if interval is FALSE}
}
\value{
A vector with (lwr, median, upr) if interval is TRUE, and a
  data.frame with prior value (x) and prior density (y)
}
\description{
Extract prior information for initial biomass, terminal U/UMSY (ustatus)
and terminal B/BMSY (bstatus) from a fit, which is useful for plotting
and other tasks.
}
\details{
This function always returns prior values evaluated in natural
space, regardless of the distribution used. Thus lognormal priors will
be converted to natural space.
}
