% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prior_posterior.R
\name{get_prior_posterior}
\alias{get_prior_posterior}
\title{plot_prior_posterior}
\usage{
get_prior_posterior(fit, driors, draws = 1000, prob = 0.9)
}
\arguments{
\item{fit}{}

\item{driors}{}

\item{draws}{}

\item{prob}{}
}
\value{
a list of prior-posterior comparisons
}
\description{
plot_prior_posterior
}
