% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tmb_model.R
\name{get_tmb_model}
\alias{get_tmb_model}
\title{Compile and load TMB models for srapls}
\usage{
get_tmb_model(model_name = "sraplus_tmb")
}
\arguments{
\item{model_name}{the name of the sraplus model to load}
}
\description{
Compile and load TMB models for srapls
}
