% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srafit.R
\name{is.srafit}
\alias{is.srafit}
\title{Test whether an object has class 'srafit'}
\usage{
is.srafit(x)
}
\arguments{
\item{x}{Object to be tested}
}
\value{
TRUE of FALSE whether it is an srafit object.
}
\description{
Test whether an object has class 'srafit'
}
