% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_draws}
\alias{plot_draws}
\title{Plot the prior draws from a fit, showing which crashed the population
and which were kept}
\usage{
plot_draws(fit)
}
\arguments{
\item{fit}{A returned list from run.SIR}
}
\value{
Nothing. A plot is created.
}
\description{
Plot the prior draws from a fit, showing which crashed the population
and which were kept
}
