% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_fit}
\alias{plot_fit}
\title{Plot management quantities for one or more fits}
\usage{
plot_fit(..., lims = c(0, 3))
}
\arguments{
\item{...}{A series of fits as returned by run.SIR}

\item{lims}{The limits for metrics. A single range can be passed and
used for all 4 metrics, or a list of ranges. NULL specifies to
determine the ranges for each metric from the data.}

\item{fit}{A returned list from run.SIR}

\item{names}{An optional character string with length equal to the
number of fits, used to identify fits on the plot.}

\item{plot}{Boolean for whether to print the plot.}
}
\value{
Functions using base plots return nothing, but those using
  ggplot2 return an invisible ggplot object which can then be used to
  print or save as normally done. In all cases if 'plot=TRUE' then a
  plot is created.
}
\description{
Plot management quantities for one or more fits
}
