% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_penalties}
\alias{plot_penalties}
\title{Plot priors vs posteriors for carrying capacity, initial depletion,
terminal year U/UMSY and B/BMSY for a series of fits.}
\usage{
plot_penalties(..., names = NULL, plot = TRUE, percentile = 0.99)
}
\arguments{
\item{...}{A series of fits as returned by run.SIR}

\item{names}{An optional character string with length equal to the
number of fits, used to identify fits on the plot.}

\item{plot}{Boolean for whether to print the plot.}

\item{percentile}{Optional percentile to calculate prior
density range. Defaults to 99\%.}

\item{fit}{A returned list from run.SIR}
}
\value{
Functions using base plots return nothing, but those using
  ggplot2 return an invisible ggplot object which can then be used to
  print or save as normally done. In all cases if 'plot=TRUE' then a
  plot is created.

An invisible ggplot2 object which can be printed or saved.
}
\description{
Plot priors vs posteriors for carrying capacity, initial depletion,
terminal year U/UMSY and B/BMSY for a series of fits.
}
