% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-prior-posterior.R
\name{plot_prior_posterior}
\alias{plot_prior_posterior}
\title{plot_prior_posterior}
\usage{
plot_prior_posterior(fit, driors, draws = 1000, prob = 0.9)
}
\arguments{
\item{fit}{}

\item{driors}{}

\item{draws}{}

\item{prob}{}
}
\value{
a plot of prior-posterior comparison
}
\description{
plot_prior_posterior
}
