% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_recdevs}
\alias{plot_recdevs}
\title{Plot the realized recruitment deviations from a fit, showing which
crashed the population and which were kept}
\usage{
plot_recdevs(fit)
}
\arguments{
\item{fit}{A returned list from run.SIR}
}
\value{
Nothing. A plot is created.
}
\description{
Plot the realized recruitment deviations from a fit, showing which
crashed the population and which were kept
}
