% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_reference}
\alias{plot_reference}
\title{Plot posterior histograms of MSY reference points and the log posterior
density}
\usage{
plot_reference(..., names = NULL, plot = TRUE)
}
\arguments{
\item{...}{A series of fits as returned by run.SIR}

\item{names}{An optional character string with length equal to the
number of fits, used to identify fits on the plot.}

\item{plot}{Boolean for whether to print the plot.}

\item{fit}{A returned list from run.SIR}
}
\value{
Functions using base plots return nothing, but those using
  ggplot2 return an invisible ggplot object which can then be used to
  print or save as normally done. In all cases if 'plot=TRUE' then a
  plot is created.
}
\description{
Plot posterior histograms of MSY reference points and the log posterior
density
}
