% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_ssb}
\alias{plot_ssb}
\title{Plot spawning biomass time series for posterior draws with catches.}
\usage{
plot_ssb(fit, ylim = NULL)
}
\arguments{
\item{fit}{A returned list from run.SIR}

\item{ylim}{The ylim to use, if not specified will be calculated
internally.}
}
\value{
Nothing. A plot is created.
}
\description{
Plot spawning biomass time series for posterior draws with catches.
}
