% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_terminal}
\alias{plot_terminal}
\title{Plot terminal year U/UMSY and B/BMSY posteriors and priors (if
applicable) for a series of fits}
\usage{
plot_terminal(..., names = NULL, plot = TRUE, xlim = NULL, q = c(0.005,
  0.995))
}
\arguments{
\item{...}{A series of fits as returned by run.SIR}

\item{names}{An optional character string with length equal to the
number of fits, used to identify fits on the plot.}

\item{plot}{Boolean for whether to print the plot.}

\item{xlim}{Optional xlim to override the defaults chosen by ggplot}

\item{q}{Optional quantiles between which to calculate prior
density. Defaults to c(0.005,0.995), that is 99\%.}
}
\value{
An invisible ggplot2 object which can be printed or saved.
}
\description{
Plot terminal year U/UMSY and B/BMSY posteriors and priors (if
applicable) for a series of fits
}
