% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_ustatus}
\alias{plot_ustatus}
\title{Plot U/UMSY (exploitation status) time series for posterior draws, with
  terminal year penalty shown if used.}
\usage{
plot_ustatus(fit, ylim = NULL)
}
\arguments{
\item{fit}{A returned list from run.SIR}

\item{ylim}{The ylim to use, if not specified will be calculated
internally.}
}
\value{
Nothing. A plot is created.
}
\description{
Plot U/UMSY (exploitation status) time series for posterior draws, with
  terminal year penalty shown if used.
}
