% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srafit.R
\name{print.srafit}
\alias{print.srafit}
\title{Print method for class 'srafit'}
\usage{
\method{print}{srafit}(fit)
}
\arguments{
\item{fit}{Object of class srafit}
}
\description{
Print method for class 'srafit'
}
