% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR.R
\name{run.SIR}
\alias{run.SIR}
\title{Run the SIR algorithm for a draws and stock (temporary)}
\usage{
run.SIR(draws, pct.keep = 10)
}
\arguments{
\item{draws}{data.frame of prior draws for SIR}
}
\value{
A list containing depletion and SSB for posterior draws, and a
  vector of Keepers
}
\description{
Run the SIR algorithm for a draws and stock (temporary)
}
