% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_species.R
\name{search_species}
\alias{search_species}
\title{Search species: Ported from https://github.com/James-Thorson-NOAA/FishLife/blob/main/R/Search_species.R}
\usage{
search_species(
  Class = "predictive",
  Order = "predictive",
  Family = "predictive",
  Genus = "predictive",
  Species = "predictive",
  add_ancestors = TRUE,
  Database = marlin::FishBase_and_RAM,
  ParentChild_gz = Database$ParentChild_gz
)
}
\arguments{
\item{Class}{Character input for taxonomic class}

\item{Order}{Character input for taxonomic class}

\item{Family}{Character input for taxonomic class}

\item{Genus}{Character input for taxonomic class}

\item{Species}{Character input for taxonomic class}

\item{add_ancestors}{Boolean whether to add ancestors for matching species or not}

\item{ParentChild_gz}{vector providing index of parent-taxon for every child-taxa}
}
\value{
integer of row numbers of \code{ParentChild_gz} matching \code{genus_species}
}
\description{
Match taxonomic inputs to a given row of \code{ParentChild_gz} or its closest ancestor
}
\details{
This function attempts to do a smart match to elements of \code{ParentChild_gz}.  It sweeps from Order to Species
and ignores any taxonomic input listed as \code{"predictive"} until it finds something else.  It then appends
\code{"predictive"} to any lower taxonomic level that is missing, and checks whether this specification yields a single,
unique taxon.  If it does, it then returns the row number and potentially any ancestors (higher taxonomic levels)
}
