% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sraplus-package.R
\docType{package}
\name{sraplus-package}
\alias{sraplus}
\alias{sraplus-package}
\title{sraplus: Run sraplus Assessments}
\description{
Runs assessments in the manner of Ovando et al 2019. Allows for SIR, Maximum LIkelihood, or HMC based fitting.
}
\keyword{internal}
