% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sraplus.R
\docType{package}
\name{sraplus}
\alias{sraplus}
\alias{sraplus-package}
\title{sraplus: Stochastic stock reduction analysis plus biological priors and
penalties}
\description{
In development, do not use!
}
\details{
This package is designed to extend the SRA analysis framework to include
  penalties on depletion and priors on life-history parameters derived
  from FishLife.
}
\references{
Thorson, J. T., S. B. Munch, J. M. Cope, and J. Gao. 2017. Predicting
life history parameters for all fishes worldwide. Ecological
Applications. 27(8):2262-2276. http://onlinelibrary.wiley.com/doi/10.1002/eap.1606/full
}
