% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sraplus_simulator.R
\name{sraplus_simulator}
\alias{sraplus_simulator}
\title{sraplus fishery simulator}
\usage{
sraplus_simulator(
  r = 0.2,
  k = 1000,
  m = 2,
  q = 1e-06,
  q_slope = 0.025,
  sigma_proc = 0.05,
  sigma_u = 0.1,
  init_dep = 1,
  init_u_umsy = 0.5,
  years = 50
)
}
\arguments{
\item{r}{}

\item{k}{}

\item{m}{}

\item{q}{}

\item{q_slope}{}

\item{sigma_proc}{}

\item{sigma_u}{}

\item{init_dep}{}

\item{init_u_umsy}{}

\item{years}{}
}
\value{

}
\description{
sraplus fishery simulator
}
