% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack-stan.R
\name{stack_stan}
\alias{stack_stan}
\title{stack reports from tmbstan fits into cohesive dataframe}
\usage{
stack_stan(stack)
}
\arguments{
\item{stack}{a stack of adreport outputs from one set of stan parameters}
}
\value{
a stacked set of parameters
}
\description{
stack reports from tmbstan fits into cohesive dataframe
}
