% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize-sraplus.R
\name{summarize_sralpus}
\alias{summarize_sralpus}
\title{provide summary of sraplus fit}
\usage{
summarize_sralpus(fit, output = "table")
}
\arguments{
\item{fit}{}

\item{output}{}
}
\value{
a summary of key sraplus outputs
}
\description{
provide summary of sraplus fit
}
