% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-sraplus.R
\name{theme_sraplus}
\alias{theme_sraplus}
\title{plotting theme for sraplus}
\usage{
theme_sraplus(base_size = 12)
}
\arguments{
\item{base_size}{the font size}
}
\value{
a theme object
}
\description{
plotting theme for sraplus
}
