% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AgeModel}
\alias{AgeModel}
\title{Stochastic SRA to calculate the trajectory given parameters}
\usage{
AgeModel(Catch, AgeMat, Steep, NatMort, AgeMax, Carry, Weight, InitialDeplete,
  Sigma)
}
\arguments{
\item{Catch}{To do}
}
\description{
Stochastic SRA to calculate the trajectory given parameters
}
