% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{draw.priors}
\alias{draw.priors}
\title{Quick function to return priors for the Mexican examples.}
\usage{
draw.priors(N, InitialDepletePrior, InitialDepleteCV, Kprior, Catch, Taxon)
}
\arguments{
\item{N}{Number of draws}

\item{InitialDepletePrior}{Mean of initial lognormal depletion penalty}

\item{InitialDepleteCV}{CV of intial lognormal depletion penalty}

\item{Kprior}{Maybe a multiplier on K?}

\item{Catch}{Vector of catch}

\item{Taxon}{A named vector with elements "Class, Order, Family, Genus,
Species". Either provide a name or "" to specify predictive.}
}
\value{
A data.frame that contains N random draws from all priors
}
\description{
Quick function to return priors for the Mexican examples.
}
