% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{draw.priors}
\alias{draw.priors}
\title{Draw samples from the prior distribution as defined by biological
  parameters from FishLife and initial depletion distribution.}
\usage{
draw.priors(N, penalties, Kprior, Catch, Taxon, Kscale = 2)
}
\arguments{
\item{N}{Number of draws}

\item{penalties}{A list of penalties, see ?run.SIR for more information.}

\item{Kprior}{Maybe a multiplier on K?}

\item{Catch}{Vector of catch}

\item{Taxon}{A list with taxonomic information, including 'Class',
'Order', 'Family', 'Genus', 'Species'. Any blanks will be converted to
'predictive' internally when querying FishLife.}

\item{Kscale}{Scalar to control the initial biomass which is generated
as runif(N, Carry/Kscale, Carry*Kscale) where Carry=Kprior*max(Catch).}
}
\value{
A list that contains 'draws', a data.frame with N random draws from all
  priors, and 'Taxon' which is a character string that is returned from
  the FishLife query match.
}
\description{
Draw samples from the prior distribution as defined by biological
  parameters from FishLife and initial depletion distribution.
}
