% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-sraplus.R
\name{fit_sraplus}
\alias{fit_sraplus}
\title{Run sraplus}
\usage{
fit_sraplus(driors, include_fit = TRUE, seed = 42, plim = 0.2,
  model = "sraplus_v2", fit_catches = TRUE,
  randos = "uc_proc_errors", draws = 1e+06, n_keep = 2000,
  use_sir = FALSE)
}
\arguments{
\item{driors}{a list of driors passed from sraplus::format_driors}

\item{include_fit}{logical indicating whether to return the fitted object}

\item{seed}{seed for model runs}

\item{plim}{cutoff (in units of B/K) for hockey stick PT function}

\item{model}{the name of the sraplus TMB version to be run}

\item{fit_catches}{logical indicating whether catches should be fit or passed}

\item{randos}{random effects when passing to TMB}

\item{draws}{the number of SIR samples to run}

\item{n_keep}{the number of SIR samples to keep}

\item{use_sir}{logical indicating whether to use SIR or TMB}
}
\value{
a fitted sraplus object
}
\description{
Run sraplus
}
