% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-driors.R
\name{format_driors}
\alias{format_driors}
\title{Format data and priors for sraplus}
\usage{
format_driors(
  taxa = "lutjanus griseus",
  initial_state = 1,
  initial_state_cv = 0.1,
  terminal_state = NA,
  terminal_state_cv = 0.2,
  u = NA,
  u_years = NA,
  u_cv = 0.2,
  terminal_u = NA,
  terminal_u_cv = NA,
  catch = NA,
  years = NA,
  index = NA,
  effort = NA,
  b_ref_type = "k",
  f_ref_type = "fmsy",
  index_years = 1,
  effort_years = NA,
  use_heuristics = FALSE,
  use_b_reg = FALSE,
  growth_rate_prior = NA,
  growth_rate_prior_cv = 0.25,
  fmi = c(research = NA, management = NA, enforcement = NA, socioeconomics = NA),
  fmi_cv = NA,
  sar = NA,
  sar_cv = NA,
  q_slope_prior = 0,
  q_slope_prior_cv = 0.1,
  m = NA,
  k_prior = NA,
  k_prior_cv = 2,
  sigma_ratio_prior = 1,
  sigma_ratio_prior_cv = 0.05,
  shape_prior = NA,
  shape_prior_cv = 0.25,
  q_prior = 0.001,
  q_prior_cv = 1,
  sigma_obs_prior = 0.05,
  sigma_obs_prior_cv = 1,
  isscaap_group = NA,
  f_prior_form = 0,
  shape_prior_source = "thorson",
  prob = 0.9,
  use_fmsy_based_r = FALSE,
  use_catch_priors = FALSE
)
}
\arguments{
\item{taxa}{the genus and species of the species (case insensitive)}

\item{initial_state}{reference point in the initial year, units of depletion or B/Bmsy (set ref_type accordingly).
when \code{initial_state} is 1, implies either B/K = 1 or B/Bmsy = 1}

\item{initial_state_cv}{CV associated with initial state reference point}

\item{terminal_state}{reference point in the terminal year, units of depletion or B/Bmsy (set ref_type accordingly).
when \code{initial_state} is 1, implies either B/K = 1 or B/Bmsy = 1}

\item{terminal_state_cv}{CV associated with terminal state reference point}

\item{u}{u/umsy data over time}

\item{u_years}{years in which u/umsy data are available}

\item{u_cv}{cv associated with u/umsy data}

\item{terminal_u}{vector of priors on u/umsy in the terminal years}

\item{terminal_u_cv}{vector of cvs on u/umsy in the terminal years}

\item{catch}{vector of catches over lifetime of fishery}

\item{years}{vector of years that the catch data correspond to}

\item{index}{vector of an abundance index}

\item{effort}{vector of an effort series}

\item{b_ref_type}{the units of the initial and final depletion. k means that units are relative to carrying capacity, b bmsy}

\item{f_ref_type}{the units of the fishing mortality rates. fmsy means F/Fmsy, f means fishing mortality rate f}

\item{index_years}{the years in which abundance index data are available}

\item{effort_years}{years in which effort data are available}

\item{use_heuristics}{logical,TRUE uses catch-msy hueristics for priors, FALSE requires user to pass them}

\item{use_b_reg}{logical marking whether to use the regression to put priors on biomass as well as fishing mortality}

\item{growth_rate_prior}{manually pass prior on the growth rate r in the Pella-Tomlinson model}

\item{growth_rate_prior_cv}{manually pass cv  for prior on the growth rate r in the Pella-Tomlinson model}

\item{fmi}{named vector of fisheries management index scrores}

\item{fmi_cv}{overwrite posterior predictive cv of FMI derived U/Umsy (NA to use empirical CV)}

\item{sar}{swept area ratio}

\item{sar_cv}{overwrite posterior predictive cv of SAR derived U/Umsy (NA to use empirical CV)}

\item{q_slope_prior}{prior on the slope of catchability q}

\item{q_slope_prior_cv}{the cv on the prior on q_slope}

\item{m}{natural mortality}

\item{k_prior}{prior on carrying capacity}

\item{k_prior_cv}{CV of prior on carrying capacity}

\item{sigma_ratio_prior_cv}{}

\item{shape_prior}{prior on shape parameter of Pella-Tomlinson model}

\item{shape_prior_cv}{CV of prior on shape parameter of Pella-Tomlinson model}

\item{q_prior_cv}{CV of prior on q itself (prior on q set in \code{fit_sraplus})}

\item{sigma_obs_prior}{prior on observation error}

\item{sigma_obs_prior_cv}{cv of prior on observation error}

\item{isscaap_group}{the isscaap group of the stock. If NA will try and find}

\item{prob}{the probability range of the posterior predictive to use in construction of the priors}
}
\value{
a list of data and priors
}
\description{
Format data and priors for sraplus
}
