% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-posterior.R
\name{get_posterior}
\alias{get_posterior}
\title{get posterior draws from tmbstan fit object}
\usage{
get_posterior(draws, inits, sra_data, model_name, randos, knockout)
}
\arguments{
\item{draws}{parameter draws}

\item{inits}{initial values for default model}

\item{sra_data}{data}

\item{randos}{random effects}

\item{knockout}{parameters to knock out}

\item{model}{model name}
}
\value{
tibble of posterior draws
}
\description{
get posterior draws from tmbstan fit object
}
