% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_fit}
\alias{plot_fit}
\title{Plot management quantities for one or more fits}
\usage{
plot_fit(..., names = NULL)
}
\arguments{
\item{...}{A series of fits of class srafit}

\item{names}{Optional vector of names}
}
\description{
Plot management quantities for one or more fits
}
