% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_fit}
\alias{plot_fit}
\title{Plot management quantities for one or more fits}
\usage{
plot_fit(..., names = NULL, lims = c(0, 3))
}
\arguments{
\item{...}{A series of fits of class srafit}

\item{names}{Optional vector of names}

\item{lims}{The limits for metrics. A single range can be passed and
used for all 4 metrics, or a list of ranges. NULL specifies to
determine the ranges for each metric from the data.}
}
\description{
Plot management quantities for one or more fits
}
