% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_reference}
\alias{plot_reference}
\title{Plot posterior histograms of MSY reference points and the log posterior
density}
\usage{
plot_reference(fit)
}
\arguments{
\item{fit}{A returned list from run.SIR}
}
\value{
Nothing. A plot is created.
}
\description{
Plot posterior histograms of MSY reference points and the log posterior
density
}
