% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-sraplus.R
\name{plot_sraplus}
\alias{plot_sraplus}
\title{Plot timeseries of estimates from multiple sraplus models}
\usage{
plot_sraplus(
  ...,
  fontsize = 14,
  years = NA,
  plotvars = c("b_div_bmsy", "c_div_msy", "depletion", "u_div_umsy"),
  kobe = FALSE,
  max_kobe_val = 4
)
}
\arguments{
\item{...}{any number of sraplus models}

\item{fontsize}{font size for plots}

\item{years}{replacement vector of years for plotting if desired}

\item{plotvars}{variables to plot}

\item{kobe}{FALSE to return ribbons plot. TRUE to plot a Kobe plot}

\item{max_kobe_val}{the max value for the x and y axes of the Kobe plot}
}
\value{
a ggplot object
}
\description{
Plot timeseries of estimates from multiple sraplus models
}
