% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-sraplus.R
\name{plot_sraplus}
\alias{plot_sraplus}
\title{Plot timeseries of estimates from multiple sraplus models}
\usage{
plot_sraplus(..., fontsize = 14, years = NA)
}
\arguments{
\item{...}{any number of sraplus models}

\item{fontsize}{font size for plots}

\item{years}{the years that the model fit applies to}
}
\value{
a ggplot object
}
\description{
Plot timeseries of estimates from multiple sraplus models
}
