% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srafit.R
\name{print.srafit}
\alias{print.srafit}
\title{Print method for class 'srafit'}
\usage{
\method{print}{srafit}(x, ...)
}
\arguments{
\item{x}{Object of class srafit}

\item{...}{Other arguments to be passed}
}
\description{
Print method for class 'srafit'
}
