% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR.R
\name{run.SIR}
\alias{run.SIR}
\title{Run the SIR algorithm for a draws and stock (temporary)}
\usage{
run.SIR(Catch, draws, deplete.mean = NULL, deplete.cv = NULL,
  deplete.distribution = 1, harvest.distribution = 1, harvest.mean = NULL,
  harvest.sd = NULL, pct.keep = 10, ProcessError = TRUE,
  penalties = NULL, simulation = NULL)
}
\arguments{
\item{Catch}{Vector of catches, one for each year.}

\item{draws}{data.frame of prior draws for SIR}

\item{deplete.mean}{Final year depletion (not in log space)}

\item{deplete.cv}{Final year CV, used as SD}

\item{deplete.distribution}{The likelhiood to use, either (1) normal or
(2) lognormal}

\item{harvest.distribution}{Same choice as deplete.distribution.}

\item{harvest.mean}{Same as harvest.sd but the mean.}

\item{harvest.sd}{The mean and SD of the terminal year
penalty on fishing pressure. If either is NULL it is ignored.}

\item{pct.keep}{The percentage of "keepers" from the total. Default 10\%.}

\item{ProcessError}{Flag for whether to include process error
deviations, passed on to model. Defaults to TRUE.}
}
\value{
A list containing depletion, SSB, and harvest rate (U) for
  posterior draws, and a vector of Keepers
  @export
}
\description{
Run the SIR algorithm for a draws and stock (temporary)
}
