% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR.R
\name{run.SIR}
\alias{run.SIR}
\title{Run the SIR algorithm for a draws and stock (temporary)}
\usage{
run.SIR(nrep, Catch, Taxon, InitialDepletePrior, InitialDepleteCV, Kprior = 3,
  deplete.mean = NULL, deplete.cv = NULL, deplete.distribution = 1,
  harvest.distribution = 1, harvest.mean = NULL, harvest.sd = NULL,
  AgeVulnOffset = -1, years = NULL, pct.keep = 10, ProcessError = TRUE,
  penalties = NULL, simulation = NULL)
}
\arguments{
\item{nrep}{Number of replicates to run.}

\item{Catch}{Vector of catches, one for each year.}

\item{Taxon}{A list with taxonomic information, including 'Class',
'Order', 'Family', 'Genus', 'Species'. Any blanks will be converted to
'predictive' internally when querying FishLife.}

\item{InitialDepletePrior}{Initial depletion mean}

\item{InitialDepleteCV}{Initial depletion CV}

\item{deplete.mean}{Final year depletion (not in log space)}

\item{deplete.cv}{Final year CV, used as SD}

\item{deplete.distribution}{The likelhiood to use, either (1) normal or
(2) lognormal}

\item{harvest.distribution}{Same choice as deplete.distribution.}

\item{harvest.mean}{Same as harvest.sd but the mean.}

\item{harvest.sd}{The mean and SD of the terminal year penalty on
fishing pressure. If either is NULL it is ignored.}

\item{AgeVulnOffset}{Optional parameter for offset of age of
vulnerability from age of maturity. A value of 0 means knife edge
selectivity starts at 50% maturity (AgeMat), while -1 means 1 year
before it. Thus (age of vulnerability = age of maturity +
AgeVulnOffset). Defaults -1. Note that AgeMat is stochastic inside the
function, so age at vulnerability will vary with it.}

\item{years}{Optional vector of years which is used by plotting
functions. Defaults to 1:length(Catch).}

\item{pct.keep}{The percentage of "keepers" from the total. Default
10\%.}

\item{ProcessError}{Flag for whether to include process error
deviations, passed on to model. Defaults to TRUE.}

\item{penalties}{A list specifying the penalties to use.}

\item{simulation}{See documentation for AgeModel.}
}
\value{
A list containing depletion, SSB, and harvest rate (U) for
  posterior draws, and a vector of Keepers
}
\description{
Run the SIR algorithm for a draws and stock (temporary)
}
