% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sraplus-package.R
\docType{package}
\name{sraplus-package}
\alias{sraplus}
\alias{sraplus-package}
\title{sraplus: Run sraplus Assessments}
\description{
Runs assessments in the manner of Ovando et al (2021). Allows for SIR, Maximum Likelihood, or HMC based fitting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://danovando.github.io/sraplus/,https://github.com/DanOvando/sraplus}
}

}
\keyword{internal}
