% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srafit.R
\name{summary.srafit}
\alias{summary.srafit}
\title{Print method for class 'srafit'}
\usage{
\method{summary}{srafit}(object, ...)
}
\arguments{
\item{object}{Object of class srafit}

\item{...}{Other arguments to be passed}
}
\description{
Print method for class 'srafit'
}
